Reorganized file structure
Later reorganization changes will not be that drastic
