+++ AppStateManager.hpp
@@ -5,7 +5,7 @@
 
 //|||||||||||||||||||||||||||||||||||||||||||||||
 
-#include "states/AppState.hpp"
+#include "scenes/AppState.hpp"
 
 //|||||||||||||||||||||||||||||||||||||||||||||||
