+++ Scenes/PerformanceScene/Bars.cpp
