+++ Scenes/PerformanceScene/Bars.h
