+++ Scenes/PerformanceScene/FingerPlacement.cpp
