+++ Scenes/PerformanceScene/Strings.cpp
