+++ Scenes/PerformanceScene/Strings.h
